requirements.txt
llama_index==0.9.40
openai==1.12.0
python-dotenv==1.0.1
